refactor: extract authOptions to its own file
